IO.py
Alright so this is sorta like a rough draft on what the functions that handle events in the db should look like. I will still do more for albums and other needed playlist as required. I just wanted to know what you think and if I am on the right path here.